refactor(discord): bootstrap の MCP/OpenCode 環境変数構築を environment.ts へ分離する
